requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
discord==1.7.3
discord.py==1.7.3
gTTS==2.2.4
idna==3.3
multidict==6.0.2
pycparser==2.21
PyNaCl==1.5.0
python-environ==0.4.54
requests==2.27.1
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.9
yarl==1.7.2
